requirements.txt
python-dotenv
jinja2
uvicorn
fastapi
starlette
requests
pillow
python-slugify
tortoise-orm
uvicorn
aerich
pytz
asyncpg

